Added directory component logic and moved makeTilesArray
