+++ src/componentLogic/makeTilesArray.js
